Refactor use admin rental hook
